+++ superwall/src/main/java/com/superwall/sdk/billing/Billing.kt
@@ -17,4 +17,6 @@ interface Billing {
     suspend fun queryAllPurchases(): List<Purchase>
 
     suspend fun consume(purchaseToken: String): Result<String>
+
+    suspend fun getStorefrontCountryCode(): String?
 }